+++ src/utils/packages/__tests__/find-package-root.js
@@ -34,6 +34,19 @@ describe('findPackageRoot', () => {
 			expect(await findPackageRoot('foo/bar/baz')).toBe('foo/bar');
 		});
 
+		it('should resolve the package directory when starting with a file path on a platform that throws ENOTDIR', async () => {
+			mockedFS.__mockFileSystem({
+				foo: {
+					bar: {
+						'package.json': '{}',
+						baz: createError('ENOTDIR', 'foo/bar/baz', -20, 'stat'),
+					},
+				},
+			});
+
+			expect(await findPackageRoot('foo/bar/baz')).toBe('foo/bar');
+		});
+
 		it("should not resolve when package.json isn't in the tree", async () => {
 			mockedFS.__mockFileSystem({
 				foo: {
@@ -82,7 +95,7 @@ describe('findPackageRoot', () => {
 			expect(await findPackageRoot('foo/bar/baz')).toBeUndefined();
 		});
 
-		it('should throw when encountering a file system error other than ENOENT', async () => {
+		it('should throw when encountering a file system error other than ENOENT or ENOTDIR', async () => {
 			mockedFS.__mockFileSystem({
 				foo: {
 					bar: {
@@ -115,6 +128,19 @@ describe('findPackageRoot', () => {
 					bar: {
 						'yarn.lock': '',
 						baz: '',
+					},
+				},
+			});
+
+			expect(await findPackageRoot('foo/bar/baz', 'yarn.lock')).toBe('foo/bar');
+		});
+
+		it('should resolve the package directory when starting with a file path on a platform that throws ENOTDIR', async () => {
+			mockedFS.__mockFileSystem({
+				foo: {
+					bar: {
+						'yarn.lock': '{}',
+						baz: createError('ENOTDIR', 'foo/bar/baz', -20, 'stat'),
 					},
 				},
 			});
@@ -170,7 +196,7 @@ describe('findPackageRoot', () => {
 			expect(await findPackageRoot('foo/bar/baz', 'yarn.lock')).toBeUndefined();
 		});
 
-		it('should throw when encountering a file system error other than ENOENT', async () => {
+		it('should throw when encountering a file system error other than ENOENT or ENOTDIR', async () => {
 			mockedFS.__mockFileSystem({
 				foo: {
 					bar: {