+++ src/utils/packages/find-package-root.js
@@ -34,7 +34,10 @@ async function findPackageRoot(startPath, rootFile = 'package.json') {
 
 			currentDirectory = parent;
 		} catch (error) {
-			if (/** @type {{code?: string}} */ (error).code === 'ENOENT') {
+			if (
+				/** @type {{code?: string}} */ (error).code === 'ENOENT' ||
+				/** @type {{code?: string}} */ (error).code === 'ENOTDIR'
+			) {
 				const parent = path.dirname(currentDirectory);
 
 				if (!path.relative(parent, currentDirectory)) {